Creates Decrypter module/test, moves tests
